// src/api/hospital/models/PerformanceEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Employee List Api
 * Hospital Employee Administration for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: xkello@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';

/**
 *
 * @export
 * @interface PerformanceEntry
 */
export interface PerformanceEntry {
    /**
     * Unique id of the performance entry
     * @type {string}
     * @memberof PerformanceEntry
     */
    id: string;
    /**
     * Type of activity (examination, surgery, preoperative consultation, checkup)
     * @type {string}
     * @memberof PerformanceEntry
     */
    activityType: string;
    /**
     * Name of the patient
     * @type {string}
     * @memberof PerformanceEntry
     */
    patientName: string;
    /**
     * Date of the activity in DD/MM/YY format
     * @type {string}
     * @memberof PerformanceEntry
     */
    activityDate: string;
    /**
     * Details of the operation (up to 255 characters)
     * @type {string}
     * @memberof PerformanceEntry
     */
    details: string;
}

/**
 * Check if a given object implements the PerformanceEntry interface.
 */
export function instanceOfPerformanceEntry(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "activityType" in value;
    isInstance = isInstance && "patientName" in value;
    isInstance = isInstance && "activityDate" in value;
    isInstance = isInstance && "details" in value;

    return isInstance;
}

export function PerformanceEntryFromJSON(json: any): PerformanceEntry {
    return PerformanceEntryFromJSONTyped(json, false);
}

export function PerformanceEntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): PerformanceEntry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        'id': json['id'],
        'activityType': json['activityType'],
        'patientName': json['patientName'],
        'activityDate': json['activityDate'],
        'details': json['details'],
    };
}

export function PerformanceEntryToJSON(value?: PerformanceEntry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        'id': value.id,
        'activityType': value.activityType,
        'patientName': value.patientName,
        'activityDate': value.activityDate,
        'details': value.details,
    };
}